Make enableRowOpacity a fn and remove the row event listener on subsequent clicks
For now, I have transfered the code that enablesRowOpacity into a function that is called
when opacityBtn is clicked and have set it to remove the event listener placed on the
rows on every click after the first one. This prevents the event listeners from stacking,
but 1 will always remain present.

For now, I am hiding the one that is present by setting the opacity to 1 in the if statement,
but I will try to see if I can completely remove it.