Make button instantly remove event listeners added to rows after 1st click
This is a temporary fix that simply removes the event listeners that would be added
on top of the 1st one after the 1st click. For now, the only way to remove the opacity
toggle is by resetting the grid. I will have to come back to this and find a solution
after more study.